+++ core-api/src/main/kotlin/com/ifmix/api/core/modules/scan/service/ScanService.kt
@@ -17,7 +17,7 @@ import com.ifmix.api.core.modules.scan.ScanRunner
 import com.ifmix.api.core.modules.scan.dto.ScanInput
 import com.ifmix.api.core.modules.scan.dto.ScanMediaItem
 import com.ifmix.api.core.modules.scan.repo.ScanRecordRepository
-import com.ifmix.api.core.model.ScanRecord
+import com.ifmix.api.core.model.scan.ScanRecord
 import org.springframework.stereotype.Service
 import java.time.Duration
 import java.time.Instant
@@ -61,7 +61,7 @@ class AntiqueService(
             appId = appId,
             imageKeys = input.images.map { ImageRef(key = it.imageKey) },
             result = result,
-            status = ScanStatus.COMPLETED.code.toShort(),
+            status = ScanStatus.COMPLETED.code,
             clientIp = ctx.clientIp,
             lang = ctx.lang,
             country = ctx.country,
